// src/main/domain/models/roll/RollType.ts
export enum RollType {
  CHAIR = 'CHAIR',
  ESPRIT = 'ESPRIT',
  ESSENCE = 'ESSENCE',
  EMPIRIQUE = 'EMPIRIQUE',
  MAGIE_LEGERE = 'MAGIE_LEGERE',
  MAGIE_FORTE = 'MAGIE_FORTE',
  SOIN = 'SOIN',
  APOTHEOSE = 'APOTHEOSE',
  ARCANE_FIXE = 'ARCANE_FIXE',
  ARCANE_ESPRIT = 'ARCANE_ESPRIT',
  ARCANE_ESSENCE = 'ARCANE_ESSENCE',
  SAUVEGARDE_VS_MORT = 'SAUVEGARDE_VS_MORT',
  RELANCE = 'RELANCE'
}